install flow and annotation feedback
